requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.8.1
    # via
    #   django
    #   django-allauth
    #   django-browser-reload
    #   django-cors-headers
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
black==25.1.0
    # via -r requirements.in
brotli==1.1.0
    # via whitenoise
certifi==2025.1.31
    # via requests
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via black
crispy-bootstrap5==2025.4
    # via -r requirements.in
distlib==0.3.9
    # via virtualenv
dj-database-url==3.0.0
    # via -r requirements.in
django==5.1.6
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
    #   dj-database-url
    #   django-allauth
    #   django-browser-reload
    #   django-cors-headers
    #   django-crispy-forms
    #   djangorestframework
    #   drf-spectacular
django-allauth==65.4.1
    # via -r requirements.in
django-browser-reload==1.18.0
    # via -r requirements.in
django-cors-headers==4.7.0
    # via -r requirements.in
django-crispy-forms==2.4
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    #   drf-spectacular
dotenv==0.9.9
    # via -r requirements.in
drf-spectacular==0.28.0
    # via -r requirements.in
dynaconf==3.2.10
    # via -r requirements.in
exceptiongroup==1.3.0
    # via pytest
filelock==3.17.0
    # via virtualenv
greenlet==3.1.1
    # via playwright
gunicorn==23.0.0
    # via -r requirements.in
identify==2.6.8
    # via pre-commit
idna==3.10
    # via requests
inflection==0.5.1
    # via drf-spectacular
iniconfig==2.0.0
    # via pytest
isort==6.0.1
    # via -r requirements.in
jsonschema==4.23.0
    # via drf-spectacular
jsonschema-specifications==2024.10.1
    # via jsonschema
mypy-extensions==1.0.0
    # via black
nodeenv==1.9.1
    # via pre-commit
packaging==24.2
    # via
    #   black
    #   gunicorn
    #   pytest
pathspec==0.12.1
    # via black
platformdirs==4.3.6
    # via
    #   black
    #   virtualenv
playwright==1.50.0
    # via
    #   -r requirements.in
    #   pytest-playwright
pluggy==1.5.0
    # via pytest
pre-commit==4.1.0
    # via -r requirements.in
psycopg2==2.9.10
    # via -r requirements.in
pyee==12.1.1
    # via playwright
pytest==8.3.4
    # via
    #   -r requirements.in
    #   pytest-base-url
    #   pytest-django
    #   pytest-playwright
pytest-base-url==2.1.0
    # via pytest-playwright
pytest-django==4.10.0
    # via -r requirements.in
pytest-playwright==0.7.0
    # via -r requirements.in
python-dotenv==1.0.1
    # via dotenv
python-slugify==8.0.4
    # via pytest-playwright
pyyaml==6.0.2
    # via
    #   drf-spectacular
    #   pre-commit
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via pytest-base-url
rpds-py==0.23.1
    # via
    #   jsonschema
    #   referencing
sqlparse==0.5.3
    # via django
text-unidecode==1.3
    # via python-slugify
tomli==2.2.1
    # via
    #   black
    #   pytest
typing-extensions==4.12.2
    # via
    #   asgiref
    #   black
    #   dj-database-url
    #   exceptiongroup
    #   pyee
    #   referencing
uritemplate==4.1.1
    # via drf-spectacular
urllib3==2.3.0
    # via requests
virtualenv==20.29.2
    # via pre-commit
whitenoise[brotli]==6.9.0
    # via -r requirements.in
